fix(suite-native): enter on trezor listens to error
