+++ suite-native/module-authorize-device/src/useRedirectOnPassphraseCompletion.ts
@@ -1,8 +1,11 @@
 import { useEffect } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import {
+    cancelPassphraseAndSelectStandardDeviceThunk,
     selectDeviceRequestedAuthorization,
+    selectHasAuthFailed,
+    selectHasVerificationCancelledError,
     selectPassphraseError,
 } from '@suite-native/device-authorization';
 import { useAuthorizationGoBack } from '@suite-native/device-authorization';
@@ -10,6 +13,10 @@ import { useAuthorizationGoBack } from '@suite-native/device-authorization';
 export const useRedirectOnPassphraseCompletion = () => {
     const hasDeviceRequestedAuthorization = useSelector(selectDeviceRequestedAuthorization);
     const hasPassphraseError = useSelector(selectPassphraseError);
+    const hasAuthorizationFailed = useSelector(selectHasAuthFailed);
+    const hasVerificationCancelledError = useSelector(selectHasVerificationCancelledError);
+
+    const dispatch = useDispatch();
 
     const { handleGoBack } = useAuthorizationGoBack();
 
@@ -19,4 +26,11 @@ export const useRedirectOnPassphraseCompletion = () => {
             handleGoBack();
         }
     }, [handleGoBack, hasDeviceRequestedAuthorization, hasPassphraseError]);
+
+    useEffect(() => {
+        // User has canceled the authorization process on device (authorizeDeviceThunk rejects with auth-failed error)
+        if (hasAuthorizationFailed || hasVerificationCancelledError) {
+            dispatch(cancelPassphraseAndSelectStandardDeviceThunk());
+        }
+    }, [dispatch, hasAuthorizationFailed, hasVerificationCancelledError]);
 };
